kms/nv50: always call drm_dp_dpcd_write after drm_dp_mst_hpd_irq
When using the dock for the Lenovo P50 laptop it could happen quite often that
on redocking the laptop no attached displays are found. Same situation when
displays are connected to the dock.

There seems to be no unhandled bits in the esi though.

With this patch detecting the displays doesn't seem to fail anymore on above
mentioned setup.

Maybe there is a better fix for this issue.

Signed-off-by: Karol Herbst <kherbst@redhat.com>